Turned fabsf into fabs and sqrtf into sqrt because those single-precision functions are not universal.
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1518 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Misc/PlayerDialogs.cpp
@@ -273,8 +273,8 @@ bool Configure_ChaseCam(ChaseCamData &Data)
 				else
 				{
 					// Overdamped case
-					float NewDampAbs = fabsf(New_Damping);
-					BadValue = ((NewDampAbs + sqrtf(-New_Spring)) >= 1);
+					float NewDampAbs = fabs(New_Damping);
+					BadValue = ((NewDampAbs + sqrt(-New_Spring)) >= 1);
 					if (BadValue)
 					{
 						float Temp = 1 - NewDampAbs;
